:recycle:  refactor: getKST utils/functions로 이동
