+++ src/pages/CreatePostPage/components/ThirdStep/index.tsx
@@ -16,6 +16,7 @@ import UploadHeader from '../UploadHeader';
 import SelectedUserList from '../../../../components/organism/SelectedUserList';
 import SearchUser from '../../../../components/organism/SearchUser';
 import { useSelectedUserList } from '../../../../contexts/SelectedUserListProvider';
+import { getKST } from '../../../../utils/functions/getKST';
 
 const ThirdStep = ({
   latitude,
@@ -55,16 +56,6 @@ const ThirdStep = ({
     const tomorrow = getKST(true).getDate();
     const storedDate = new Date(date).getDate();
     return tomorrow > storedDate ? false : true;
-  };
-
-  const getKST = (tomorrow: boolean) => {
-    const curr = new Date();
-    const utc = curr.getTime() + curr.getTimezoneOffset() * 60 * 1000;
-    const KR_TIME_DIFF = 9 * 60 * 60 * 1000;
-    const TOMORROW_TIME_DIFF = 24 * 60 * 60 * 1000;
-    const kr_curr = new Date(utc + KR_TIME_DIFF);
-    const kr_tomorrow = new Date(utc + KR_TIME_DIFF + TOMORROW_TIME_DIFF);
-    return tomorrow ? kr_tomorrow : kr_curr;
   };
 
   useEffect(() => {
